Refactor country code handling and validation scripts
- Updated `countryCodes` in `EditCustomer.cshtml` and `NewCustomer.cshtml` to use numeric keys for country codes.
- Added `_ValidationScriptsPartial` to both views for improved form validation.
- Rearranged script inclusions in `_Layout.cshtml` for better loading order and added comments for clarity.
- Corrected script references in `_ValidationScriptsPartial.cshtml` and fixed a typo in the unobtrusive validation path.
- Introduced a nullable `Gender` property in `EditCustomerViewModel` for gender selection.
- Modified `GetGenderList` and `GetCountryList` in `CustomerService.cs` to use integer values for `SelectListItem`.

+++ Services/Customer/CustomerService.cs
@@ -142,7 +142,8 @@ namespace Services.Customer
             var Genders = Enum.GetValues<Gender>()
                 .Select(g => new SelectListItem
                 {
-                    Value = g.ToString(),
+                    Value = ((int)g).ToString(),
+                    //Value = g.ToString(),
                     Text = g.ToString()
                 })
                 .ToList();
@@ -153,7 +154,8 @@ namespace Services.Customer
             var Countries = Enum.GetValues<Country>()
                 .Select(c => new SelectListItem
                 {
-                    Value = c.ToString(),
+                    Value = ((int)c).ToString(),
+                    //Value = c.ToString(),
                     Text = c.ToString()
                 })
                 .ToList();
